sql-edu-backend/requirements.txt
# FastAPI 核心
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库
sqlalchemy>=2.0.0
aiomysql>=0.2.0
alembic>=1.12.0

# 数据验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# bcrypt 4.1+ 与 passlib 的版本探测兼容性有问题，固定到 4.0.1
bcrypt==4.0.1

# AI 服务
openai>=1.0.0

# 邮件服务
fastapi-mail>=1.3.0
aiosmtplib>=2.0.0

# CORS
python-multipart>=0.0.6

# 测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
# AI 生成题目/翻译等功能在某些环境下通过 SOCKS 代理访问大模型；
# httpx 需要 socksio 才能支持 SOCKS 代理，否则会报 ImportError。
socksio>=1.0.0
aiosqlite>=0.19.0  # SQLite 异步驱动，用于测试数据库

pyjwt